fix 32bit warnings, use strncat/cpy instead

+++ src/sentry_scope.c
@@ -175,13 +175,13 @@ sentry__symbolize_frame(const sentry_frame_info_t *info, void *data)
         && sentry_value_is_null(
             sentry_value_get_by_key(frame, "symbol_addr"))) {
         sentry_value_set_by_key(frame, "symbol_addr",
-            sentry__value_new_addr((uint64_t)info->symbol_addr));
+            sentry__value_new_addr((uint64_t)(intptr_t)info->symbol_addr));
     }
 
     if (info->load_addr
         && sentry_value_is_null(sentry_value_get_by_key(frame, "image_addr"))) {
         sentry_value_set_by_key(frame, "image_addr",
-            sentry__value_new_addr((uint64_t)info->load_addr));
+            sentry__value_new_addr((uint64_t)(intptr_t)info->load_addr));
     }
 }
 
@@ -209,7 +209,8 @@ sentry__symbolize_stacktrace(sentry_value_t stacktrace)
         if (!addr) {
             continue;
         }
-        sentry__symbolize((void *)addr, sentry__symbolize_frame, &frame);
+        sentry__symbolize(
+            (void *)(intptr_t)addr, sentry__symbolize_frame, &frame);
     }
 }
